moved nix types to file they are being used in, renamed files
